Move containers to container folder (Header, SideBar, Logout, ReduxBreadcrumb)
