--- src/components/SideBarConfig/index.js
@@ -1,3 +1,0 @@
-/* @flow */
-import SideBarConfig from './SideBarConfig'
-export default SideBarConfig
