+++ src/components/index.js
@@ -5,18 +5,14 @@
 export { default as BranchSelect } from './BranchSelect'
 export { default as DropDown } from './DropDown/DropDown'
 export { default as ErrorMessage } from './ErrorMessage'
-export { default as Logout } from './Logout'
 export { default as NotificationList } from './NotificationList/NotificationList'
 export { default as ProjectList } from './ProjectList/ProjectList'
-export { default as SideBar } from './SideBar/SideBar'
-export { default as SideBarConfig } from './SideBarConfig/SideBarConfig'
 export { default as Tree } from './Tree/Tree'
 export { default as CodeMirrorView } from './CodeMirrorView/CodeMirrorView'
 export { default as TextEditorBox } from './TextEditorBox/TextEditorBox'
 export { default as UserFilter } from './UserFilter/UserFilter'
 export { default as Filter } from './Filter/Filter'
 export { default as Checkbox } from './Checkbox/Checkbox'
-export { default as ReduxBreadcrumb } from './ReduxBreadcrumb/ReduxBreadcrumb'
 export { default as SelectableList } from './SelectableList/SelectableList'
 export { default as Link } from './Link/Link'
 export { default as LinkButton } from './LinkButton/LinkButton'
@@ -26,7 +22,6 @@ export { default as Label } from './Label/Label'
 export { default as CodeDiffView } from './CodeDiffView/CodeDiffView'
 export { default as Comment } from './Comment/Comment'
 export { default as NewComment } from './NewComment/NewComment'
-export { default as Header } from './Header'
 export { default as UserAvatar } from './UserAvatar/UserAvatar'
 export { default as TestAvatar } from './TestAvatar/TestAvatar'
 export { default as SearchBox } from './SearchBox/SearchBox'
